Fit the dungeon viewport to the terminal window (#4)
* Fit the dungeon viewport to the terminal window

The game always drew into a fixed 80x24 corner, leaving the rest of a
larger terminal blank. The dungeon viewport now auto-fits the actual
terminal size by default (up to a full 66x198 level), with a -W
COLSxLINES CLI override.

SCREEN_HEIGHT/SCREEN_WIDTH/MAX_HEIGHT/MAX_WIDTH (dungeon generation
topology and the save file's implicit-length floor grid) are untouched,
so gameplay and save compatibility are unaffected. A new, independent
viewport size governs only how much of the already-generated level is
shown at once: at the classic 80x24 size the panel-follow logic is
byte-identical to before; at any larger size it centers on the player
and clamps to the level bounds instead of stepping through the classic
discrete half-screen panel positions.


* Fix enlarged-viewport panel getting stuck at zero size

coord_outside_panel_enlarged()'s "skip redraw if unchanged" checks
compared the freshly computed top/left against dg().panel's current
values, but a newly generated level resets panel.top/bottom/left/right
to 0 (generate_cave()) before this ever runs. For the town — and any
dungeon level where the player starts near the origin — the correctly
computed top/left also lands on 0, so the checks read as "unchanged"
and skipped ever writing panel.bottom/right/col_prt, leaving them at
their stale zeroed-out values from the reset.

That left col_prt at 0 instead of the correct left-13, which put
freshly drawn dungeon tiles directly into the sidebar's columns, and
left panel.bottom/right at 0, so only a single row/column of the level
was ever drawn — matching the reported symptoms: the town not fully
visible, a blank panel after leaving a store (draw_cave_panel() does a
full redraw, but of whatever bogus bounds were left over), and dungeon
tiles overwriting the stats sidebar.

Fix: also require the panel's current size to already match the live
viewport before trusting it as "unchanged".


* Center the town/dungeon view and static screens in the terminal

Two related fixes, both leaving MAX_HEIGHT/MAX_WIDTH/SCREEN_HEIGHT/
SCREEN_WIDTH and the save format untouched:

- The town (and any dungeon level smaller than the viewport) was always
  drawn pinned to the viewport's top-left corner, dwarfed by a much
  bigger auto-fit viewport on modern terminals. coord_outside_panel_enlarged()
  now centers the level's actual content within the viewport (equal
  margin on all sides) via row_prt/col_prt, instead of just clamping
  top/left to 0.

- Static full-page screens (splash, character creation, store, help,
  death, high scores) were always pinned to the terminal's absolute
  top-left, even though the live dungeon view already auto-fits the
  terminal. Added a small offset system in ui_io.rs: every drawing
  primitive adds one of two fixed, precomputed offsets (dungeon-view
  footprint vs. classic 80x24) before the real curses call, and each
  static screen opts into the classic offset via center_for_static_screen().
  The default stays the dungeon-view offset (usually a no-op) so a
  screen nobody marked doesn't render mis-aligned, just un-centered.
  Screens reachable from both a static and dungeon-view context (help)
  save/restore whichever was active rather than assuming.

wizard.rs was audited and needs no changes: none of its functions
display a full-page screen, they're all inline prompts over whatever
is already showing.


* Center the viewport on the player, not the map

Entering a level used to clamp the view to the map's bounds, so a player
who spawned near a map corner appeared near the screen corner — leaking
their position within the level. The enlarged-viewport path now centers
an unclamped conceptual window on the player: the stored panel stays the
window/map intersection (so every consumer indexing the floor by panel
bounds remains in-bounds), while row_prt/col_prt place the window on
screen, leaving blank space past the map's edges. The player lands on
the exact center cell at every level entry and re-center; the town's
content-centering special case is gone in favor of the same rule.

draw_dungeon_panel now erases the whole viewport (not just the clamped
panel rows) so a repositioned window can't leave stale tiles behind.

Adds a curses-free integration test pinning the centering, hysteresis,
and map-intersection invariants.


* Double the dungeon to 132x396 and scale level content to match

MAX_HEIGHT/MAX_WIDTH double in each dimension (4x area); the town stays
22x66. Per-level content scales x4 to keep upstream density: rooms
(DUN_ROOMS_MEAN 32->128), streamers, stairs, monsters, objects, gold,
and traps, with normal-distribution std-devs scaled in lockstep.
Capacity limits rise to match (MON_TOTAL_ALLOCATIONS 125->500,
LEVEL_MAX_OBJECTS 175->700), which forced Tile.creature_id/treasure_id
from u8 to u16. Painting counts stay as-is: their save encoding can't
fit a x4 target.

Save-format fallout: x coordinates no longer fit a byte, so monster,
painting, and sparse floor-record coordinates widen to shorts (the
floor records' terminator becomes 0xFFFF). The save file now carries a
separate save-format version (6.0.0) so old saves are rejected with the
friendly message instead of being misparsed; the displayed game version
stays 5.7.15, and the high-score file format is untouched.

The 'M' overview map picks the finest shrink ratio (2, 3, or 6) that
fits the current viewport instead of a fixed 3, falling back to the
classic 22x66 footprint. Also fixes command_locate_on_map's y-bounds
check, which compared against the map width (an upstream bug whose
effect the wider map would have doubled).


* Reword save/score version-mismatch messages to name jdbkmoria

These errors surfaced the upstream "umoria" name even though the
save-format check is jdbkmoria's own, and a player loading a
completely unrelated file would get a misleading message.


* Clamp monster distance to u8 to fix divide-by-zero on the doubled map

`distance_from_player` is a u8 assigned from `coord_distance_between()`
(an i32). After the dungeon was doubled to 132x396, distances reach ~460,
so the `as u8` cast truncated mod 256 -- landing on exactly 0 at distances
of 256/512 and panicking in `100 / distance` (monster.rs), and corrupting
sight/distance checks in between. Clamp to 255 before the cast; every
threshold the field is compared against is <=20, so this is
gameplay-equivalent and leaves the save format (still a byte) unchanged.


* Stop running when passing a visible painting

Running (find mode) now ends when the run scan sees a painting on an
adjacent wall, under the same visibility gate as the doorway stop --
paintings are landmarks, not part of the corridor.


* Guarantee the map painting is easy to find on levels 1-2

On dungeon levels 1 and 2 the level-map painting is now the painting
nearest the character's arrival position among those hanging on a
lit-room wall, so it is always visible and close by. If no painting
happened to land on a lit-room wall, one more is hung on the nearest
such wall; a level with no lit room at all falls back to the old rule.
Levels 3+ keep the random designation.


* Hang the level 1-2 map painting on the lit wall nearest the start

"Nearest of the ~50 randomly scattered paintings" typically landed ~20
tiles away in another room on the doubled map, so the map painting was
still hard to find. Now levels 1-2 scan the whole level for the
lit-room hangable wall tile nearest the character's arrival position
and put the map there -- designating the painting already on that wall
or hanging a fresh one. Median distance from the start is now 2 tiles.


---------

+++ src/dungeon_tile.rs
@@ -6,8 +6,13 @@
 // Tile holds data about a specific tile in the dungeon.
 #[derive(Debug, Clone, Copy, Default)]
 pub struct Tile {
-    pub creature_id: u8, // ID for any creature occupying the tile
-    pub treasure_id: u8, // ID for any treasure item occupying the tile
+    // jdbkmoria extension: widened from u8 to u16. LEVEL_MAX_OBJECTS (700)
+    // and MON_TOTAL_ALLOCATIONS (500) - the capacities these fields index
+    // into - were scaled x4 alongside the dungeon's area and both now
+    // exceed u8::MAX, so a u8 here would silently wrap. See game_save.rs
+    // for the matching save-format widening (wr_short/rd_short).
+    pub creature_id: u16, // ID for any creature occupying the tile
+    pub treasure_id: u16, // ID for any treasure item occupying the tile
     pub feature_id: u8,  // ID of cave feature; walls, floors, open space, etc.
 
     pub perma_lit_room: bool,  // Room should be lit with perm light, walls with this set should be perm lit after tunneled out.
